Implements fl slider dim (not done)
The flashlight_dim_layer.png is still wrong, is just like a "proof of concept", i just need to get a more accurate value for it's oppacity. the way it works is that when a slider is holded it renders a dark sprite with some level of oppacity in it. just doing the pull request now so people can review it.