Reduce audio delay/latency to a known sane minimum
Note that this increases CPU usage, but not to a concerning state. Additionally, after the BASS library update that was shipped with 1.6.8 3rd pre-release, players claim the audio becomes more offset, and the low latency synchronization setting became what it swore to destroy.

Since the system allocates less resources to background applications, some adjustments had to be made to BASS configurations when the game is acting as a background music player, otherwise the playback will be choppy. This has a caveat of having to reinitialize BASS and reload the music every time the player tabs in and out of the game, but it is better than having to delete the music player altogether.