Rewrite beatmap parser and difficulty calculation flow
Integrating the new difficulty calculation is a pain with the current structure of the beatmap parser. Additionally, the beatmap parser is due to a rewrite, anyway. As such, I have rewritten the beatmap parser in a way that should be easier to work with.

This also replaces `DifficultyReCalculator` with a new and structurally better `BeatmapDifficultyCalculator`.