Revert unranked sprite persistence logic to semi-old behavior
Having it persist in the center throughout gameplay is an anti-design. However, I remember that the sprite was changed to persist permanently as many players did not realize that their play was unranked.

To find a good middle ground, I allowed the sprite to exist until the first object appears, with a cap of 1.5 seconds. In addition to that, the sprite expands as it disappears, giving more signal to the player.