Convert `ScrollableList` to a `UIComponent`
This allows old components that still rely on `ScrollableList` to take advantage of `UIComponent` features until `ScrollableList` itself is removed.

It also fixes `ScrollableList`s not receiving inputs as it relied on `Scene` input handling, which requires `ITouchArea`s to be registered directly into the `Scene` (meaning in a nested `Scene` scenario, it would not work without using `childScene`).