Update HUD element toolbar anchor and origin instead of position to prevent out of bounds
This somehow fixes a problem where the buttons in the toolbar feels difficult to press.

During investigation, I found out that the reported local position in the button can vary (but would still occasionally work due to input binding), likely due to a race condition between the draw and main thread. This can be a sign that the bug lies somewhere else, but due to time constraints, this fix will do for the time being.