Remove `BeatmapDifficulty` parameter from `Mod.calculateScoreMultiplier`
This was only used in `ModDifficultyAdjust`, and it already has a reference from each statistics' default values. In addition to that, `StatisticV2` use cases where score multiplier needs to be calculated is only during gameplay (which already has a complete `Beatmap` reference). As such, the parameter is safe to remove.