Revert "Unload textures that have the same key in `ResourceManager`"
This reverts commit 20ead05a72e95353cd0167802b5db2f4506275e8.

Consider the following scenario:
- Upon launching, the logged in player's avatar is loaded to the engine's texture manager via the main menu's `OnlinePanel` (let's call the loaded texture `tex1`)
- Going to song selection menu or multiplayer lobby reloads the avatar, which creates a new texture (let's call it `tex2`) and unloads `tex1` from the engine's texture manager because of this change
- Main menu's `OnlinePanel` still uses `tex1`, which has been unloaded by the engine and therefore displays a white rectangle

While this does not fully represent all cases since avatar handling is quite convoluted, this shows that the change breaks reused components that do not reload its textures.

A full fix to this problem takes significant effort, so for the time being, a revert is enough.