Continuously update rank texture in break period
Closes https://github.com/osudroid/osu-droid/issues/204.

This is caused by an object that is missed after the break period is initiated due to hit window leniency. A "real" fix would be hard to do with the current code, but continuous texture region update does not incur a high performance cost, so I figured that this is an acceptable solution.