Never "unload" beatmap filter in song select menu
For whatever reason, the unload operation makes the fragment `null`, which can cause an NPE when any operation involving it (e.g., sorting and reloading cards) is not done.

Ideally, these operations should be aborted if the sorting mode is changed, but this is going to take more effort than a simple change like this.